remove LWRP, custom resource
